I transferred the GmailAPI to the service package and renamed it to ResumeService
